Rename address-watch files back to agent
